Some Standardizing of variables and declarations
All attributes made private and declared at the top of the class.
If you want X and Y to be doubles, your constructors should take
doubles.
Based on your use descripton, I would say a float is better.
WIth private attributes you can always be sure that this.attribute
always refers to the attribute in this class and not any passed in
parameters.